requirements.txt
Django>=4.2
mysqlclient
gunicorn
python-decouple
djangorestframework
whitenoise
Pillow